Refactor PatientCard.java
